ONE-9: Fix clang build errors

+++ vnet/vnet/lisp-gpe/lisp_gpe.c
@@ -198,8 +198,8 @@ add_del_negative_fwd_entry (lisp_gpe_main_t * lgm,
       /* TODO insert tunnel that always sends map-request */
     case DROP:
       /* for drop fwd entries, just add route, no need to add encap tunnel */
-      adj.lookup_next_index =  ip_prefix_version(dpref) == IP4 ?
-              LGPE_IP4_LOOKUP_NEXT_DROP : LGPE_IP6_LOOKUP_NEXT_DROP;
+      adj.lookup_next_index =  (u32) (ip_prefix_version(dpref) == IP4 ?
+              LGPE_IP4_LOOKUP_NEXT_DROP : LGPE_IP6_LOOKUP_NEXT_DROP);
 
       /* add/delete route for prefix */
       return ip_sd_fib_add_del_route (lgm, dpref, spref, a->table_id, &adj,
